# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backend API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/groups": {
            "get": {
                "tags": [
                    "Group"
                ],
                "summary": "GroupsInfo",
                "operationId": "42bae3f26bf83e45b919fcbd6f2720d7",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "groups": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-12-20 17:49:36"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-12-20 17:49:36"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "OAIp2"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Group"
                ],
                "summary": "Create",
                "operationId": "89db403173427de890a5d41ca8519a08",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "OAIp2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "group": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "OAIp2"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/group/{group}/user/{user}": {
            "post": {
                "tags": [
                    "Group User"
                ],
                "summary": "Create",
                "operationId": "b5aa6ca7cd0462798c4b508e275a8ad5",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Group ID",
                        "required": true,
                        "example": 1
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "group": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "OAIp2"
                                                },
                                                "users": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "first_name": {
                                                                "type": "string",
                                                                "example": "User"
                                                            },
                                                            "last_name": {
                                                                "type": "string",
                                                                "example": "Userov"
                                                            },
                                                            "middle_name": {
                                                                "type": "string",
                                                                "example": "Userovich"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Group User"
                ],
                "summary": "Delete",
                "operationId": "c960d7b87ccf75b4b23abfa0a033591b",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Group ID",
                        "required": true,
                        "example": 1
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/groups/{group}": {
            "get": {
                "tags": [
                    "Group"
                ],
                "summary": "GetGroupID",
                "operationId": "b77150428df0b8d1c3af9552e763dc65",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Group ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "group": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "OAIp2"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Group"
                ],
                "summary": "Update",
                "operationId": "331761998f3b0aa1a801eb01d3f4fdaa",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Group ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "OAIp2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "group": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-12-20 17:49:36"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "OAIp2"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Group"
                ],
                "summary": "Delete",
                "operationId": "54094f01a1f0294a063e25e282917d39",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Group ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/group/users": {
            "get": {
                "tags": [
                    "Group User"
                ],
                "summary": "GroupsUserInfo",
                "operationId": "fe1aa7d093536f4d78394cdabd4ac00e",
                "parameters": [
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Role name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "student",
                                "teacher"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "groups": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-12-20 17:49:36"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-12-20 17:49:36"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "OAIp2"
                                                    },
                                                    "users": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "first_name": {
                                                                    "type": "string",
                                                                    "example": "User"
                                                                },
                                                                "last_name": {
                                                                    "type": "string",
                                                                    "example": "Userov"
                                                                },
                                                                "middle_name": {
                                                                    "type": "string",
                                                                    "example": "Userovich"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}